UI Initial Setup
Set up the player HUD and new color selection menu.
